% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accused_witches.R
\docType{data}
\name{accused_witches}
\alias{accused_witches}
\title{Accused witches}
\format{
A data frame with 152 rows and 5 variables:
\describe{
\item{Accused.Witch}{Name of accused}
\item{Residence}{Town of residence}
\item{Month.of.Accusation}{Month of accusation}
\item{Month.of.Execution}{Month of execution}
\item{Sort}{ID from original data set}
}
}
\source{
\url{https://www.tulane.edu/~salem}
}
\usage{
accused_witches
}
\description{
A dataset containing tax list of accused witches.
}
\keyword{datasets}
