% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cent17th.R
\docType{data}
\name{cent17th}
\alias{cent17th}
\title{Names and link segments}
\format{
An object of class \code{data.frame} with 156 rows and 17 columns.
}
\source{
\url{http://17thc.us/primarysources/accused.php?id=1&pg=1}
}
\usage{
cent17th
}
\description{
A dataset based on data about the accused found
http://17thc.us/primarysources/accused.php?id=1&pg=1
}
\details{
containing a comprehensive list of names from the records
if the Salem Witchcraft Trials and the correspondent stub indicating
the page name associated with them.

\describe{
\item{table}{Source table}
\item{Name}{Name of those who were condemned during the trial}
\item{Marital Statues (if female) (source)}{The marital status of those condemned if they were female with the source}
\item{1. evidence of foreiture (source) 2. Article seized}{The evidence they had for forefeiture, the source in which they had and if they had aritcle seized or not for those who were condemned}
\item{Legal statues (source)}{their legal status in which this case is condmemned and their souce}
\item{Attainder reversed}{if the declaration of the attainders were reversed or not for the condemned}
\item{Comments}{Any comments made for those who were condemned}
}
}
\keyword{datasets}
