% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Forfeitures_Description.R
\docType{data}
\name{forfeitures}
\alias{forfeitures}
\title{forfeitures
A data set containing list of forfeitures for those who were condemned, escaped,
and stoodcmute during trial. Name, Marital Status (If married) and source,
\enumerate{
\item evidence of forfeiture (source) 2. Article seized, Legal statues (source),
Attainder reversed, and Comments.
}}
\format{
A data frame with 30 rows and 6 variables with the name Person_condemned:
\describe{
\item{table}{Source table}
\item{Name}{Name of those who were condemned during the trial}
\item{Marital Statues (if female) (source)}{The marital status of those condemned if they were female with the source}
\item{1. evidence of foreiture (source) 2. Article seized}{The evidence they had for forefeiture, the source in which they had and if they had aritcle seized or not for those who were condemned}
\item{Legal statues (source)}{their legal status in which this case is condmemned and their souce}
\item{Attainder reversed}{if the declaration of the attainders were reversed or not for the condemned}
\item{Comments}{Any comments made for those who were condemned}
}

The source has some Key Tables which are:
\itemize{
\item for the 1703 and 1711 acts reversing some of the witchcraft attainder see Commonwealth of Massachusetts, the Act and Resolver, Public and Private, of the providence of the Massachusetts Bay, 21 vols. (Boston, 1869-1922), VI, 49-50 (1703 act), VI, 71-72 (1711 act).
SWP-Paul Boyer and Stephen Nissenbain, eds., The Salem Witchcraft Papers; Verbian Transcript of the Legal Documents of the Salem Witchcraft Outbreak of 1692, 3 vols. (New York, 1977).
Calef-Robert Calef, More Wonders of the Invisible World (London, 1700), excerpted in George L. Burr, ed., Narratives of the Witchcraft Cases, 1684-1706 (New York, 1914), 293-393.
Brattle-Thomas Brattle, "Letter," Oct. 8, 1692, in Burr, ed., Narratives, 169-190.
Hale-John Hale, A modest Inquiry Into the Nature of Witchcraft (Boston, 1702), in Burr, Narratives, 399-432.
Gage-Thomas Gage, the History of Rowley (Boston, 1840).
}
}
\source{
{Brown, David C. “The Forfeitures at Salem, 1692.” The William and Mary Quarterly,#' vol. 50, no. 1, 1993, pp. 85–111. JSTOR, www.jstor.org/stable/2947237?seq=1, 10.2307/2947237}
}
\usage{
forfeitures
}
\description{
forfeitures
A data set containing list of forfeitures for those who were condemned, escaped,
and stoodcmute during trial. Name, Marital Status (If married) and source,
\enumerate{
\item evidence of forfeiture (source) 2. Article seized, Legal statues (source),
Attainder reversed, and Comments.
}
}
\keyword{datasets}
