% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/long_doc_name.R
\docType{data}
\name{long_doc_name}
\alias{long_doc_name}
\title{Names and link segments}
\format{
An object of class \code{data.frame} with 3884 rows and 2 columns.
}
\source{
\url{https://salem.lib.virginia.edu/category/swp.html}
}
\usage{
long_doc_name
}
\description{
A dataset containing a comprehensive list of names mentioned in therecords
of the Salem Witchcraft Papers and the correspondent stub indicating
the page name associated with them.
}
\details{
There are 139 separate pages which mention a total of 1392 unique names.
857 individuals were mentioned once. The maximum number of mentions is 85
(John Hathorn, who served as Salem's Magistrate).

\describe{
\item{link}{Source table}
\item{name}{Name of person mentioned}
}
}
\keyword{datasets}
