% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parris_social.R
\docType{data}
\name{parris_social}
\alias{parris_social}
\title{The social positions of pro-Parris and anti-Parris individuals}
\format{
A data frame with 189 rows and 5 variables:
\describe{
\item{Name}{Name of the individual}
\item{view}{Stance on Parris}
\item{Sex}{Sex of the individual}
\item{Identification}{The social position of the individual (five groups)}
\item{Sort}{ID from original data set}
}
}
\source{
\url{https://www.tulane.edu/~salem/Pro-Parris\%20Data\%20Set.html}
}
\usage{
parris_social
}
\description{
A dataset containing a list of individuals who signed pro-Parris
and anti-Parris petitions and their social positions
}
\keyword{datasets}
