% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salem_village.R
\docType{data}
\name{salem_village}
\alias{salem_village}
\title{Accused witches}
\format{
A data frame with 152 rows and 5 variables:
\describe{
\item{Name}{Name of taxpayer (may be a family)}
\item{Petition}{Did this person sign either the pro- or anti-Parris petition in 1695}
\item{Church.to.1696}{Was the person a member of any church through 1696}
\item{Sort}{ID from original data set}
}
}
\source{
\url{https://www.tulane.edu/~salem}
}
\usage{
salem_village
}
\description{
A dataset containing tax list of accused witches.
}
\keyword{datasets}
