% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_comparison.R
\docType{data}
\name{tax_comparison}
\alias{tax_comparison}
\title{Tax list data fo Salem}
\format{
A data frame with 200 rows and 9 variables:
\describe{
\item{Name}{Name of the tax payer}
\item{Petition}{Stance on Petition}
\item{Year}{Year of payment}
\item{Tax}{Tax paid}
\item{Sort}{ID from original data set}
}
}
\source{
\url{https://www.tulane.edu/~salem}
}
\usage{
tax_comparison
}
\description{
A dataset containing tax list data for Salem, MA, 1681-1700.
Payments are in shillings (converted from original records).
Petition refers to Pro-Parris, Anti-Parris and No Signature.
Data are in long format with one record per person per year.
}
\keyword{datasets}
