% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tax_comparison.R
\docType{data}
\name{tax_comparison}
\alias{tax_comparison}
\title{Tax list data fo Salem}
\format{
A data frame with 200 rows and 9 variables:
\describe{
\item{Name}{Name of the tax payer}
\item{Petition}{Stance on Petition}
\item{X1681}{Payment in 1681}
\item{X1690}{Payment in 1690}
\item{X1694}{Payment in 1694}
\item{X1695}{Payment in 1695}
\item{X1697}{Payment in 1697}
\item{X1700}{Payment in 1700}
\item{Sort}{ID from original data set}
}
}
\source{
\url{https://www.tulane.edu/~salem}
}
\usage{
tax_comparison
}
\description{
A dataset containing tax list data for Salem, MA, 1681-1700.
Payments are in shillings (converted from original records).
Petition refers to Pro-Parris, Anti-Parris and No Signature
}
\keyword{datasets}
